Remove validate and mutate webhooks from v1beta1 MySQLCluster.
